Refactor model imports for consistency and update import paths
- Updated import statements across multiple controllers and services to use the new model file names, ensuring consistent casing and improving code clarity.
- Removed deprecated model files to streamline the codebase and prevent potential issues on case-sensitive file systems.
