+++ bot/controller/agent/agentCrawlerController.js
@@ -1,5 +1,5 @@
 import AgentCrawlerService from '../../services/nlp/AgentCrawlerService.js';
-import Agent from '../../models/agent.js';
+import Agent from '../../models/agentModel.js';
 
 class AgentCrawlerController {
   /**
