+++ bot/controller/user/userController.js
@@ -1,4 +1,4 @@
-import User from '../../models/User.js';
+import User from '../../models/userModel.js';
 import BaseController from '../shared/baseController.js';
 
 /**
