+++ bot/models/agentVectorModel.js
