+++ bot/models/companyModel.js
