+++ bot/models/knowledgeModel.js
