+++ bot/services/CompanyService.js
@@ -1,6 +1,6 @@
 import crypto from 'crypto';
-import Company from '../models/Company.js';
-import User from '../models/User.js';
+import Company from '../models/companyModel.js';
+import User from '../models/userModel.js';
 
 class CompanyService {
   constructor() {
